backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
conda==4.3.16
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
distlib==0.3.4
Django==2.2.19
django-environ==0.8.1
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.4.0
pipenv==2022.1.8
platformdirs==2.5.0
psycopg2-binary==2.8.6
pycosat==0.6.3
pycparser==2.21
python3-openid==3.2.0
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
virtualenv-clone==0.5.7
weasyprint==52.5
zipp==3.6.0
